refactor: extract task configs into registry
- Create frontier_swe_env/tasks/ package with registry pattern
- Move PG instruction + factory functions to tasks/pg.py
- task_config.py now only has TaskConfig class + backward-compat
re-exports
- Environment __init__ accepts task_name/mode params, resolves via
registry
- from frontier_swe_env.tasks import get_task_config, list_tasks
- All old import paths still work (PG_TRAINING_INSTRUCTION,
pg_training_config)
